split FastGradientMethod into separate module
